fix profilePage and MainLayout
