+++ src/components/Header/Header.jsx
@@ -9,7 +9,15 @@ import RecoveryPasswordForm from '../RecoveryPasswordForm/RecoveryPasswordForm';
 import { CustomLink } from '../CustomLink/CustomLink';
 import logotype from '../../assets/icons/logotype.svg';
 
-export const Header = observer(({ onLogin, textError, loggedIn, signOut }) => {
+/////////Авторизация//////
+import {
+  postApiAutorisation,
+  postApiAuthorizeRefresh,
+  postApiAuthorizeVerify,
+} from '../../utils/Auth.js';
+/////////////////////////
+
+export const Header = observer(() => {
   // временное решение, пока не реализовано апи авторизации
   // const [loggedIn, setLoggedIn] = useState(true);
   const location = useLocation();
@@ -18,6 +26,76 @@ export const Header = observer(({ onLogin, textError, loggedIn, signOut }) => {
     const element = document.getElementById(id);
     element.scrollIntoView({ behavior: 'smooth', block: 'nearest' });
   };
+
+    /////////Авторизация//////
+    const [loggedIn, setLoggedIn] = useState(false);
+    const [textError, setTextError] = useState('');
+    console.log(loggedIn);
+  
+    const handleLogin = (email, password) => {
+      return postApiAutorisation(email, password)
+        .then((data) => {
+          //в (data) должны прийти два токена access и refresh
+          //токен обновления refresh
+          const refresh = data.refresh;
+  
+          if (data.access) {
+            //если получили токен доступа access
+            //токен доступа access в куки, максимальное время хранения 1 час (max-age=3600)?
+            document.cookie = `access=${data.access}; max-age=3600`;
+  
+            //делаем стейт залогиненного
+            setLoggedIn(true);
+            //и закрываем попап формы
+            formStore.setOpenAuthForm(false);
+          } else {
+            //если не получили токен доступа access
+            //отправляем токен обновления на сервер
+            //проверяем действителен ли токен обновления
+            postApiAuthorizeVerify(refresh)
+              .then((data) => {
+                if (data) {
+                  //если токен обновления действителен, отправляем токен обновления на сервер
+                  /////Принимает веб-токен JSON типа обновления и возвращает веб-токен JSON типа доступа
+                  postApiAuthorizeRefresh(refresh)
+                    .then((data) => {
+                      // получили токен доступа access
+                      if (data.access) {
+                        //если получили токен доступа access
+                        //токен доступа access в куки, максимальное время хранения 1 час (max-age=3600)?
+                        document.cookie = `access=${data.access}; max-age=3600`;
+  
+                        //делаем стейт залогиненного
+                        setLoggedIn(true);
+                        //и закрываем попап формы
+                        formStore.setOpenAuthForm(false);
+                      }
+                    })
+                    .catch((err) => {
+                      console.log(err, '--authorizeRefresh,err');
+                      setLoggedIn(false);
+                      setTextError(err.message);
+                    });
+                }
+              })
+              .catch((err) => {
+                console.log(err, '--token,err');
+                setLoggedIn(false);
+                setTextError(err.message);
+              });
+          }
+        })
+        .catch((err) => {
+          console.log(err, '--authorize,err');
+          setLoggedIn(false);
+          setTextError(err.message);
+        });
+    };
+
+    const handleSignOut = () => {
+      setLoggedIn(false);
+    };
+    /////////////////////////
 
   return (
     <header className={styles.header}>
@@ -69,7 +147,7 @@ export const Header = observer(({ onLogin, textError, loggedIn, signOut }) => {
           {location.pathname === '/profile' && (
             <span className={styles.email}>user@email.ru</span>
           )}
-          {!loggedIn && (
+          {!loggedIn ? (
             <button
               className={styles.button_header}
               type="button"
@@ -77,12 +155,11 @@ export const Header = observer(({ onLogin, textError, loggedIn, signOut }) => {
             >
               Войти
             </button>
-          )}
-          {loggedIn && (
+          ) : (
             <button
               className={styles.button_header}
               type="button"
-              onClick={signOut}
+              onClick={handleSignOut}
             >
               Выйти
             </button>
@@ -93,7 +170,7 @@ export const Header = observer(({ onLogin, textError, loggedIn, signOut }) => {
           >
             {formStore.showAuthForm && (
               <AuthTabs
-                onLogin={onLogin}
+                onLogin={handleLogin}
                 loggedIn={loggedIn}
                 textError={textError}
               />
