+++ src/stores/index.js
@@ -1,5 +1,7 @@
 import FormStore from './form-store';
+import AuthSore from './auth-store';
 
 const FormGlobalStore = new FormStore();
+const AuthGlobalStore = new AuthSore();
 
-export { FormGlobalStore };
+export { FormGlobalStore, AuthGlobalStore };
